Categories select list in admin panel fixed
Added dynamically listed categories wiht possibility to change category of each video
