+++ src/Controller/Admin/SuperAdmin/CategoriesController.php
@@ -5,6 +5,7 @@ namespace App\Controller\Admin\SuperAdmin;
 use App\Entity\Category;
 use App\Form\CategoryType;
 use App\Utils\CategoryTreeAdminList;
+use App\Utils\CategoryTreeAdminOptionList;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -90,5 +91,16 @@ class CategoriesController extends AbstractController
             return true;
         }
         return false;
+    }
+
+    public function getAllCategories(CategoryTreeAdminOptionList $categories, $editedCategory = null): Response
+    {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
+        $categories->getCategoryList($categories->buildTree());
+        return $this->render('admin/_all_categories.html.twig', [
+            'categories' => $categories,
+            'editedCategory' => $editedCategory
+        ]);
     }
 }
