+++ templates/admin/edit_category.html.twig
@@ -3,12 +3,12 @@
 {% block mainadmin %}
 <h2>Editing category</h2>
 
-{{ form_start(form, {'attr':{'novalidate':'novalidate'}}) }}
+{{ form_start(form) }}
     {{ form_widget(form.name, {'attr':{'class':'form-control' ~ is_invalid, 'placeholder':'Category name'}, 'id':'validationServer01'})}}
     <br>
     <label for="inlineFormCustomSelect">Parent:</label>
     <select name="category[parent]" class="custom-select mr-sm-2" id="inlineFormCustomSelect">
-        {{ render(controller('App\\Controller\\Admin\\MainController::getAllCategories', {'editedCategory': category})) }}
+        {{ render(controller('App\\Controller\\Admin\\SuperAdmin\\CategoriesController::getAllCategories', {'editedCategory': category})) }}
     </select>
     <div class="invalid-feedback">
         {{ form_errors(form.name) }}
